+++ functions/rollup.config.js
@@ -1,3 +1,5 @@
+import resolve from 'rollup-plugin-node-resolve';
+
 /**
  * Add here external dependencies that actually you use.
  */
@@ -9,6 +11,7 @@ const externals = [
 export default {
     input: 'tmp/index.js',
     external: externals,
+    plugins: [resolve()],
     output: {
         file: 'lib/index.js',
         format: 'cjs',