+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Firebase Functions TypeScript starter Changelog
 
+<a name="Apr 4, 2018"></a>
+### Apr 4, 2018
+* Update to Cloud Functions for Firebase SDK v1.0
+
 <a name="Jan 15, 2018"></a>
 ### Jan 15, 2018
 * Create project